backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Database
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
psycopg2-binary==2.9.9

# Auth
PyJWT==2.8.0
cryptography==41.0.7
authlib==1.3.0
requests==2.31.0

# WebAuthn (biometric login)
py-webauthn==2.0.0

# Security
flask-limiter==3.5.0
flask-talisman==1.1.0

# 2FA
pyotp==2.9.0
qrcode==7.4.2
